+++ api/src/admin/guards/roles.guard.ts
@@ -2,30 +2,28 @@ import {
   Injectable,
   CanActivate,
   ExecutionContext,
-  Inject,
-  forwardRef,
   UnauthorizedException,
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { User } from 'src/user/models/user.interface';
-import { UserService } from 'src/user/service/user.service';
+import { RetrievalService } from 'src/user/retrieval/retrieval.service';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
   constructor(
     private reflector: Reflector,
-    @Inject(forwardRef(() => UserService))
-    private userService: UserService,
+    private retrievalService: RetrievalService,
   ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
     const roles = this.reflector.get('roles', context.getHandler());
+    console.log('iam ran');
     if (!roles) {
       return true;
     }
     const request = context.switchToHttp().getRequest();
     const user: User = request.user;
-    const latestRolesUpdate = await this.userService.findOne(user.id);
+    const latestRolesUpdate = await this.retrievalService.findById(user.id);
     if (!roles.includes(latestRolesUpdate.role)) {
       throw new UnauthorizedException(
         'You are not allowed to access this route',
