+++ api/src/auth/auth.module.ts
@@ -1,16 +1,14 @@
-import { Module, forwardRef } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
 import { AuthService } from './service/auth.service';
 import { AuthController } from './controller/auth.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtAuthGuard } from './guards/jwt-auth.guard';
 import { JwtStrategy } from './guards/jwt.strategy';
-import { RolesGuard } from './guards/roles.guard';
 import { UserModule } from 'src/user/user.module';
 
 @Module({
   imports: [
-    forwardRef(() => UserModule),
     JwtModule.registerAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
@@ -19,8 +17,9 @@ import { UserModule } from 'src/user/user.module';
         signOptions: { expiresIn: '10000s' },
       }),
     }),
+    forwardRef(() => UserModule),
   ],
-  providers: [AuthService, JwtAuthGuard, JwtStrategy, RolesGuard],
+  providers: [AuthService, JwtAuthGuard, JwtStrategy],
   controllers: [AuthController],
   exports: [AuthService],
 })
