+++ api/src/auth/controller/auth.controller.ts
@@ -1,4 +1,12 @@
-import { Controller } from '@nestjs/common';
+import { Body, Controller, Post } from '@nestjs/common';
+import { AuthService } from '../service/auth.service';
+import { LoginDto } from '../dto/login.dto';
 
 @Controller('auth')
-export class AuthController {}
+export class AuthController {
+  constructor(private authService: AuthService) {}
+  @Post('/login')
+  login(@Body() user: LoginDto): Promise<object> {
+    return this.authService.login(user);
+  }
+}
