+++ api/src/blog/blog.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BlogEntity } from './models/blogs.entity';
 import { AuthModule } from 'src/auth/auth.module';
-import { BlogController } from './controller/blog.controller';
-import { BlogService } from './service/blog.service';
+import { BlogController } from './blog.controller';
+import { BlogService } from './blog.service';
 import { UserIsAuthorGuard } from './guards/user-is-author.guard';
 import { UserModule } from 'src/user/user.module';
 
@@ -11,5 +11,6 @@ import { UserModule } from 'src/user/user.module';
   imports: [TypeOrmModule.forFeature([BlogEntity]), AuthModule, UserModule],
   providers: [BlogService, UserIsAuthorGuard],
   controllers: [BlogController],
+  exports: [BlogService],
 })
 export class BlogModule {}
