+++ api/src/user/models/user.interface.ts
@@ -4,8 +4,8 @@ export interface User {
   id?: number;
   name?: string;
   username?: string;
-  email?: string;
-  password?: string;
+  email: string;
+  password: string;
   role?: UserRole;
   blogs?: Blog[];
   profileImg?: string;
