+++ api/src/user/user.controller.ts
@@ -1,0 +1,33 @@
+import {
+  Body,
+  Controller,
+  Param,
+  ParseIntPipe,
+  Post,
+  Put,
+  UseGuards,
+} from '@nestjs/common';
+import { UserService } from './user.service';
+import { User } from './models/user.interface';
+import { RegisterDto } from '../user/dto/register.dto';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
+import { UserValidateGuard } from 'src/user/guards/user-vaidate.guard';
+
+@Controller('user')
+export class UserController {
+  constructor(private userService: UserService) {}
+
+  @Post('/register')
+  create(@Body() user: RegisterDto): Promise<RegisterDto> {
+    return this.userService.create(user);
+  }
+
+  @UseGuards(JwtAuthGuard, UserValidateGuard)
+  @Put(':id')
+  updateOne(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() user: User,
+  ): Promise<any> {
+    return this.userService.updateOne(id, user);
+  }
+}
